feat(subway) BaseDto 생성
