+++ src/main/java/nextstep/subway/dto/StationResponse.java
@@ -1,27 +1,25 @@
 package nextstep.subway.dto;
 
+import java.time.LocalDateTime;
 import nextstep.subway.domain.Station;
 
-import java.time.LocalDateTime;
-
-public class StationResponse {
+public class StationResponse extends BaseDto {
     private Long id;
     private String name;
-    private LocalDateTime createdDate;
-    private LocalDateTime modifiedDate;
-
-    public static StationResponse of(Station station) {
-        return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(), station.getModifiedDate());
-    }
 
     public StationResponse() {
+        super();
     }
 
-    public StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+    private StationResponse(Long id, String name, LocalDateTime createdDate, LocalDateTime modifiedDate) {
+        super(createdDate, modifiedDate);
         this.id = id;
         this.name = name;
-        this.createdDate = createdDate;
-        this.modifiedDate = modifiedDate;
+    }
+
+    public static StationResponse of(Station station) {
+        return new StationResponse(station.getId(), station.getName(), station.getCreatedDate(),
+                station.getModifiedDate());
     }
 
     public Long getId() {
@@ -30,13 +28,5 @@ public class StationResponse {
 
     public String getName() {
         return name;
-    }
-
-    public LocalDateTime getCreatedDate() {
-        return createdDate;
-    }
-
-    public LocalDateTime getModifiedDate() {
-        return modifiedDate;
     }
 }
